mtp-carry: satisfy the file-size cap and rustfmt
Two CI failures on this branch, both mine.

`cargo fmt --check`: the struct literals in the new StoreRange tests were
hand-wrapped. Reformatted; the assertions are on integers and struct fields,
not on `\`-continued message strings, so nothing rustfmt reflowed could change
what they assert — re-run to confirm, 671 pass.

`Enforce ≤500 LoC`: `mtp_carry.rs` went 398 → 878 under the new docs and tests.
Split the test module into `mtp_carry_tests.rs` and include it with `#[path]`,
the pattern `toolcall_tests.rs` and `bfcl_tests.rs` already use — 462 and 426
lines, all 24 tests still pass.

`meta.rs` was the other violation and the easier one to miss: it sat at exactly
500 on main and my ticket-draw comment pushed it to 512. It is not on the
allow-list, and adding it there would need a rationale and a tracking issue for
twelve lines of comment. The rationale belonged on `StoreRange` and
`stamped_merge` anyway — keeping it in both places was duplication — so the
comment here is now a two-line pointer and the file is back to 500.
